+++ commands/u_33cm.js
@@ -1,4 +1,5 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
+const { randFromArray } = require('../lib/fnHelper');
 
 const thirtythreecm = [
 	'marry me!'
@@ -8,6 +9,6 @@ module.exports = {
 		.setName('33cm')
 		.setDescription('33cm\'s special'),
 	async execute(interaction) {
-		await interaction.reply( `${thirtythreecm[ Math.floor( Math.random() * thirtythreecm.length ) ] }` );
+		await interaction.reply( `${randFromArray(thirtythreecm)}` );
 	},
 };